[refactor] #151 add manager classes in domain package
