+++ src/main/java/org/kkumulkkum/server/domain/meeting/manager/MeetingSaver.java
@@ -1,4 +1,4 @@
-package org.kkumulkkum.server.service.meeting;
+package org.kkumulkkum.server.domain.meeting.manager;
 
 import lombok.RequiredArgsConstructor;
 import org.kkumulkkum.server.domain.meeting.Meeting;
