+++ src/main/java/org/kkumulkkum/server/domain/participant/manager/ParticipantRetriever.java
@@ -1,9 +1,9 @@
-package org.kkumulkkum.server.service.participant;
+package org.kkumulkkum.server.domain.participant.manager;
 
 import lombok.RequiredArgsConstructor;
 import org.kkumulkkum.server.domain.participant.Participant;
-import org.kkumulkkum.server.dto.participant.ParticipantStatusUserInfoDto;
-import org.kkumulkkum.server.dto.participant.response.LateComerDto;
+import org.kkumulkkum.server.api.participant.dto.ParticipantStatusUserInfoDto;
+import org.kkumulkkum.server.api.participant.dto.response.LateComerDto;
 import org.kkumulkkum.server.exception.ParticipantException;
 import org.kkumulkkum.server.exception.code.ParticipantErrorCode;
 import org.kkumulkkum.server.domain.participant.repository.ParticipantRepository;
