+++ src/main/java/org/kkumulkkum/server/domain/userinfo/manager/UserInfoRemover.java
@@ -1,18 +1,16 @@
-package org.kkumulkkum.server.service.userInfo;
+package org.kkumulkkum.server.domain.userinfo.manager;
 
 import lombok.RequiredArgsConstructor;
-import org.kkumulkkum.server.domain.userinfo.UserInfo;
 import org.kkumulkkum.server.domain.userinfo.repository.UserInfoRepository;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UserInfoSaver {
+public class UserInfoRemover {
 
     private final UserInfoRepository userInfoRepository;
 
-    public void save(final UserInfo user){
-        userInfoRepository.save(user);
+    public void deleteByUserId(final Long userId) {
+        userInfoRepository.deleteByUserId(userId);
     }
-
 }
