+++ src/main/java/org/kkumulkkum/server/domain/userinfo/manager/UserInfoRetriever.java
@@ -1,4 +1,4 @@
-package org.kkumulkkum.server.service.userInfo;
+package org.kkumulkkum.server.domain.userinfo.manager;
 
 import lombok.RequiredArgsConstructor;
 import org.kkumulkkum.server.domain.userinfo.UserInfo;
